# move Node.__lt__ SumNode special case to SumNode.__lt__