@tinyb0t 2023-09-09
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/shape/symbolic.py                15              18
------------------------------------------------------------
lines added in the tinygrad folder: -3
```