@chaosagent 2023-09-09
The speedup is entirely from less isinstance? Wow!